coldmine.go에 있는 코드를 역할에 따른 파일들로 분리
