+++ http_git.go
@@ -1,0 +1,197 @@
+package main
+
+import (
+	"encoding/base64"
+	"errors"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"log"
+	"net/http"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"time"
+)
+
+func getHead(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	headerNoCache(w)
+	sendFile(w, r, "text/plain", pth)
+}
+
+func getInfoRefs(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	r.ParseForm()
+	s := r.Form.Get("service")
+	if s == "git-upload-pack" || s == "git-receive-pack" {
+		// smart protocol
+		args := []string{"upload-pack", "--stateless-rpc", "--advertise-refs", filepath.Join(repoRoot, repo)}
+		if s == "git-receive-pack" {
+			args = []string{"receive-pack", "--stateless-rpc", "--advertise-refs", filepath.Join(repoRoot, repo)}
+		}
+		out, err := exec.Command("git", args...).CombinedOutput()
+		if err != nil {
+			log.Printf("(%v) %s", err, out)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		headerNoCache(w)
+		w.Header().Set("Content-Type", "application/x-"+s+"-advertisement")
+		p, err := packetLine("# service=" + s + "\n")
+		if err != nil {
+			log.Print(err)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		w.Write([]byte(p))
+		w.Write([]byte("0000")) // flushing
+		w.Write(out)
+	} else {
+		// dumb protocol
+		err := exec.Command("git", "update-server-info").Run()
+		if err != nil {
+			log.Print(err)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		headerNoCache(w)
+		sendFile(w, r, "text/plain", pth)
+	}
+}
+
+// packetLine adds 4 digit hex length string to given string.
+func packetLine(l string) (string, error) {
+	h := strconv.FormatInt(int64(len(l)+4), 16)
+	if len(h) > 4 {
+		return "", errors.New("packet too long")
+	}
+	return strings.Repeat("0", 4-len(h)) + h + l, nil
+}
+
+func getTextFile(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	headerNoCache(w)
+	sendFile(w, r, "text/plain", pth)
+}
+
+func getInfoPacks(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	// TODO: pack file validation.
+	headerNoCache(w)
+	sendFile(w, r, "text/plain; charset=utf-8", pth)
+}
+
+func getLooseObject(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	headerCacheForever(w)
+	sendFile(w, r, "x-git-loose-object", pth)
+}
+
+func getPackFile(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	headerCacheForever(w)
+	sendFile(w, r, "x-git-packed-objects", pth)
+}
+
+func getIdxFile(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	headerCacheForever(w)
+	sendFile(w, r, "x-git-packed-objects-toc", pth)
+}
+
+func serviceUpload(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	service(w, r, "upload-pack", repo, pth)
+}
+
+func serviceReceive(w http.ResponseWriter, r *http.Request, repo, pth string) {
+	if !checkAuth(r) {
+		w.Header().Set("WWW-Authenticate", `Basic realm="COLDMINE"`)
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte("401 Unathorized\n"))
+		return
+	}
+	service(w, r, "receive-pack", repo, pth)
+}
+
+func checkAuth(r *http.Request) bool {
+	r.ParseForm()
+	authHeader := r.Header.Get("Authorization")
+	auths := strings.Split(authHeader, " ")
+	if len(auths) != 2 || auths[0] != "Basic" {
+		return false
+	}
+	b, err := base64.StdEncoding.DecodeString(auths[1])
+	if err != nil {
+		return false
+	}
+	pair := strings.Split(string(b), ":")
+	if len(pair) != 2 {
+		return false
+	}
+	user, passwd := pair[0], pair[1]
+	if user != "coldmine" || passwd != password {
+		return false
+	}
+	return true
+}
+
+func service(w http.ResponseWriter, r *http.Request, s, repo, pth string) {
+	w.Header().Set("Content-Type", "application/x-git-"+s+"-result")
+
+	cmd := exec.Command("git", s, "--stateless-rpc", filepath.Join(repoRoot, repo))
+
+	in, err := cmd.StdinPipe()
+	if err != nil {
+		log.Print(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	out, err := cmd.StdoutPipe()
+	if err != nil {
+		log.Print(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	err = cmd.Start()
+	if err != nil {
+		log.Print(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	body, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Print(err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+	in.Write(body)
+	io.Copy(w, out)
+	cmd.Wait()
+}
+
+func headerNoCache(w http.ResponseWriter) {
+	w.Header().Set("Expires", "Fri, 01 Jan 1980 00:00:00 GMT")
+	w.Header().Set("Pragma", "no-cache")
+	w.Header().Set("Cache-Control", "no-cache, max-age=0, must-revalidate")
+}
+
+func headerCacheForever(w http.ResponseWriter) {
+	now := time.Now().Unix()
+	w.Header().Set("Date", fmt.Sprintf("%v", now))
+	w.Header().Set("Expires", fmt.Sprintf("%v", now+31536000))
+	w.Header().Set("Cache-Control", "public, max-age=31536000")
+}
+
+func sendFile(w http.ResponseWriter, r *http.Request, typ string, pth string) {
+	f, err := os.Stat(pth)
+	if err != nil {
+		if os.IsNotExist(err) {
+			w.WriteHeader(http.StatusNotFound)
+			return
+		} else {
+			log.Print(err)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	}
+	w.Header().Set("Content-Type", typ)
+	w.Header().Set("Content-Length", fmt.Sprintf("%v", f.Size()))
+	w.Header().Set("Last-Modified", f.ModTime().Format(http.TimeFormat))
+	http.ServeFile(w, r, pth)
+}
